Fix bug with process.nextTick referencing the wrong function.
`lolex` and by extension `sinon` fake timers, change the value of
`process.nextTick` in order to properly be added to the fake event loop stack.
The previous way this module was written kept a reference to the original
function and didn't look up the current value of `process.nextTick`. This
commit makes each invocation of nextTick look up the current value of
`process.nextTick`.